requirements.txt
asgiref==3.7.2
black==23.3.0
certifi==2023.5.7
chardet==3.0.4
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.8
Django==4.2.2
djangorestframework==3.14.0
djlint==1.31.0
EditorConfig==0.12.3
flake8==6.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==2.7
jsbeautifier==1.14.8
json5==0.9.14
marvel==0.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
pycodestyle==2.10.0
pyflakes==3.0.1
PyYAML==6.0
regex==2023.6.3
requests==2.20.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
urllib3==1.24.3
